pyproject.toml
[project]
name = "yunote"
version = "0.1.0"
description = "AI-powered meeting & lecture note assistant"
readme = "README.md"
license = { text = "GPL-3.0" }
authors = [{ name = "XiaoYu" }]
requires-python = ">=3.10,<3.13"
keywords = ["note", "asr", "llm", "meeting", "lecture", "summary"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: X11 Applications :: Qt",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business",
]
dependencies = [
    "requests>=2.32.4",
    "openai>=1.97.1",
    "diskcache>=5.6.3",
    "PyQt5==5.15.11",
    "PyQt-Fluent-Widgets==1.8.4",
    "modelscope>=1.32.0",
    "psutil>=7.0.0",
    "json-repair>=0.49.0",
    "pydub>=0.25.1",
    "tenacity>=8.2.0",
    "httpx[socks]>=0.28.1",
    "socksio>=1.0.0",
    "python-docx>=1.1.2",
]

[project.urls]
Homepage = "https://github.com/XiaoYu/YuNote"
Repository = "https://github.com/XiaoYu/YuNote"

[project.scripts]
yunote = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.uv]
# 为不同平台分别解析依赖（PyQt5-Qt5 版本因平台而异）
environments = [
    "sys_platform == 'win32'",
    "sys_platform == 'darwin'",
    "sys_platform == 'linux'",
]
# 覆盖 PyQt5-Qt5 版本：Windows 用 5.15.2，其他平台用最新版
override-dependencies = [
    "PyQt5-Qt5==5.15.2; sys_platform == 'win32'",
    "PyQt5-Qt5>=5.15.11; sys_platform != 'win32'",
]
dev-dependencies = [
    "pyright>=1.1.0",
    "ruff>=0.4.0",
    "pytest>=8.0.0",
]

[tool.pyright]
venvPath = "."
venv = ".venv"
pythonVersion = "3.12"
typeCheckingMode = "basic"
include = ["app", "main.py"]
exclude = [
    "**/__pycache__",
    "**/.pytest_cache",
    ".venv",
    "venv",
    "build",
    "dist",
    "work-dir",
    "AppData",
    "resource",
    "**/node_modules"
]
# 导入相关
reportMissingImports = "warning"
reportMissingTypeStubs = false
# 类型检查级别（降低严格度）
reportGeneralTypeIssues = false
reportOptionalOperand = "warning"
reportOptionalMemberAccess = false
reportArgumentType = "warning"
# 禁用的检查
reportCallIssue = false
reportUnknownMemberType = false
reportUnknownArgumentType = false
reportUnknownVariableType = false
reportUnknownParameterType = false
reportUnusedImport = "warning"
reportUnusedVariable = "warning"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --strict-markers --tb=short --disable-warnings"
markers = [
    "integration: Integration tests that require external services",
    "slow: Slow running tests",
    "llm: Tests that require LLM API access",
    "translator: Tests for translation modules",
]
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]
